feat: text is now ellipsed
