+++ frontend/src/container/TraceDetail/SelectedSpanDetails/index.tsx
@@ -1,9 +1,11 @@
-import { Tabs, Tooltip, Typography } from 'antd';
+import { Modal, Tabs, Tooltip, Typography } from 'antd';
+import Editor from 'components/Editor';
 import { StyledSpace } from 'components/Styled';
 import useThemeMode from 'hooks/useThemeMode';
-import React, { useMemo } from 'react';
+import React, { useMemo, useState } from 'react';
 import { ITraceTree } from 'types/api/trace/getTraceItem';
 
+import EllipsedButton from './EllipsedButton';
 import ErrorTag from './ErrorTag';
 import {
 	CardContainer,
@@ -25,6 +27,17 @@ function SelectedSpanDetails(props: SelectedSpanDetailsProps): JSX.Element {
 	const OverLayComponentServiceName = useMemo(() => tree?.serviceName, [
 		tree?.serviceName,
 	]);
+
+	const [isOpen, setIsOpen] = useState(false);
+
+	const [text, setText] = useState({
+		text: '',
+		subText: '',
+	});
+
+	const onToggleHandler = (state: boolean): void => {
+		setIsOpen(state);
+	};
 
 	if (!tree) {
 		return <div />;
@@ -52,18 +65,57 @@ function SelectedSpanDetails(props: SelectedSpanDetailsProps): JSX.Element {
 				</Tooltip>
 			</StyledSpace>
 
+			<Modal
+				onCancel={(): void => onToggleHandler(false)}
+				title="Log Message"
+				visible={isOpen}
+				destroyOnClose
+				footer={[]}
+				width="70vw"
+			>
+				<CustomSubTitle>{text.text}</CustomSubTitle>
+
+				{text.text === 'exception.stacktrace' ? (
+					<Editor onChange={(): void => {}} readOnly value={text.subText} />
+				) : (
+					<CustomSubText ellipsis={false} isDarkMode={isDarkMode}>
+						{text.subText}
+					</CustomSubText>
+				)}
+			</Modal>
+
 			<Tabs defaultActiveKey="1">
 				<TabPane tab="Tags" key="1">
 					{tags.length !== 0 ? (
 						tags.map((tags) => {
+							const value = tags.key === 'error' ? 'true' : tags.value;
+							const isEllipsed = value.length > 24;
+
 							return (
 								<React.Fragment key={JSON.stringify(tags)}>
 									{tags.value && (
 										<>
 											<CustomSubTitle>{tags.key}</CustomSubTitle>
-											<CustomSubText isDarkMode={isDarkMode}>
-												{tags.key === 'error' ? 'true' : tags.value}
-											</CustomSubText>
+											<Tooltip overlay={(): string => value}>
+												<CustomSubText
+													ellipsis={{
+														rows: 1,
+													}}
+													isDarkMode={isDarkMode}
+												>
+													{value}
+												</CustomSubText>
+											</Tooltip>
+											{isEllipsed && (
+												<EllipsedButton
+													{...{
+														event: tags.key,
+														onToggleHandler,
+														setText,
+														value,
+													}}
+												/>
+											)}
 										</>
 									)}
 								</React.Fragment>
@@ -75,7 +127,11 @@ function SelectedSpanDetails(props: SelectedSpanDetailsProps): JSX.Element {
 				</TabPane>
 				<TabPane tab="Events" key="2">
 					{tree.event && Object.keys(tree.event).length !== 0 ? (
-						<ErrorTag event={tree.event} />
+						<ErrorTag
+							onToggleHandler={onToggleHandler}
+							setText={setText}
+							event={tree.event}
+						/>
 					) : (
 						<Typography>No events data in selected span</Typography>
 					)}
